Fix show SAML session expiration modal
The SAML session expiration modal was not showing because `setTimeout`
was experiencing great delays when the window/tab was inactive.

In order to avoid these delays, we check regularly, every
10 seconds, against a timestamp for the expiration. Although,
`setInterval` experiences identical delays, because of the high
frequency of execution, we circumvent this issue.

In addition, we register an listener for the `visibilitychange` event,
that fires when window/tab become active.

EE: true

Changelog: fixed