Resolve StrongParams rubocop TODOs for some admin controllers
Many admin controllers use `params` directly without using Rails'
`StrongParameters` helper. This causes Rubocop exceptions, which were
ignored when the rubocop was introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/155661

This commit addresses offenses in some of the admin controllers. In
simple cases, we call `permit` directly in-line. Where the parameters
are being accessed in multiple locations, a helper method is introduced.

Resolves https://gitlab.com/gitlab-org/gitlab/-/issues/477392