Rename nightly Vue 3 Jest rule definition
For https://gitlab.com/gitlab-org/gitlab/-/issues/468194, we want to add
a new set of jobs that run Jest specs with Vue 3 in merge requests.

This makes it clearer that the hidden job definition is specifically for
the nightly pipeline run, and not just for Vue 3 Jest jobs in general.