Allow environment alerts to display alerts from any source
Alerts for an environment associated with a prometheus alert would
previously be displayed from /path/to/project/-/environments.
PrometheusAlert records specifically are no longer available, but
alerts from prometheus can still be displayed.

With this change, any alert from 'Monitor > Alerts' that is
associated with the environment will be displayed from the list.

Context:
The Metrics Dashboard (used for monitoring apps deploying using gitlab)
was deprecated in 14.7 and removed from the product in 16.0.
This commit cleans up a portion of the dead code.

https://gitlab.com/groups/gitlab-org/-/epics/10107

Changelog: changed