Added todos_vue_application feature flag
Instead of interweaving the old todos app on `/dashboard/todos` with the
new Vue app, we create a new route `/dashboard/todos/vue` which is a
little more green field.

This will allow us to work without the CSS and JS tech-debt of the old
page. Once the implementation of `/dashboard/todos/vue` is a little more
progressed we can feature flag it on the route level. But in the mean
time it's also nice that people which have the FF enabled, can just
navigate to the old route as well.

Follow-up steps would be switching out the route based on the FF in the
nav as well.

Changelog: added