Add migrations for import placeholder user limits
Adds four tiers of plan limits to limit the number of import placeholder
users that can be created per root namespace during an import, and
sets the limit values for GitLab.com plans.

Which of the four tiers of limits that apply will be determined by seat
count.

https://gitlab.com/gitlab-org/gitlab/-/issues/455903

Changelog: added