Allow Import to check placeholder loading progress
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/156704 introduced
the batch loading of import placeholder references from Redis to
PostgreSQL.

This change allows an import to check if there are values in the Redis
set.

The worker that finalises an import will include
`Import::WaitsForPlaceholderReferences`, and specify a `import_source`,
and `import_uid` that must match the values passed to
`Import::PlaceholderReferences::PushService`.

Before finishing the import the worker can call `#pending_references?`
and choose to requeue.

Here's an example of checking the progress, and retrying the finalising
import worker again later (we may as well queue another
`LoadPlaceholderContributionsWorker`):

```ruby
class MyWorker
  include ApplicationWorker
  include ::Import::WaitsForPlaceholderReferences

  REQUEUE_DELAY = 10.seconds

  import_source :example_source
  import_uid -> { project.id }

  def perform(project_id)
    @project = Project.find(project_id)

    if pending_references?
      info(
        message: 'Re-enqueuing, waiting for placeholder references',
        pending_references_count: pending_references_count
      )

      Gitlab::Import::LoadPlaceholderContributionsWorker
        .perform_async(:example_source, project.id)

      return self.class.perform_in(REQUEUE_DELAY, project.id)
    end

    # Do import finalization
  end

  private

  attr_reader :project
end
```

https://gitlab.com/gitlab-org/gitlab/-/issues/467511