Decouple scoped variables from the job record on pipeline creation
When creating a new pipeline we use rules to evaluate if a job can be
a part of the pipeline. To do this, we need to have access to variables.
And, variables could only be generated through a `Ci::Build` object.

In this MR, we are decoupling scoped variables from `Ci::Build` objects
in the pipeline creation process. With that, we expect more performant
variables calculation.

These changes are behind
the FF `ci_variables_optimization_for_yaml_and_node`.