Fix triage-ops seeds
Before this commit, anybody trying to contribute to triage-ops
will be blocked, because the tool setup has been broken for
a year (see https://gitlab.com/gitlab-org/quality/triage-ops/-/issues/1434#note_1992934408 for an example)

We don't really need to have the security project setup
for working on triage-ops. I prefer to have a working setup that
is incomplete rather than a complete one that's broken.

If anybody needs to work on the security project locally, it might
prove easier to try to add it than fixing it altogether.