Remove incorrect duplicate index entry
This index are not actually duplicates of each other because
index_unarchived_occurrences_on_version_id_and_traversal_ids is a
partial index and index_sbom_occurrences_on_component_version_id is not.
This entry was added by accident because the former index was changed to
partial during code review.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/150476#note_1896509171