Add the ability to revoke Feed Tokens via the Group revocation API
The Group Token Revocation API allows you to pass certain types of
tokens and have them revoked IF the token gives access to the group,
a subgroup, or a project within the group.

This commit adds the ability to revoke one type of Feed Token - the
long lived user-scoped feed token. (It does not allow revocation of the
path-based user feed token.)

There isn't actually a token record for Feed Tokens, they're a column on
User. (Compare with PersonalAccessTokens or  DeployTokens, which are
their own model.) Feed tokens therefore can't technically be revoked,
but instead we can rotate them and not inform the caller of the new
value. In this way we have made the old token inoperative.

If the token is successfully revoked, the API returns a minimal amount
of information about the User whose feed token it was. This response
is quite different to the response given when a PAT or DeployToken are
revoked, but that makes sense since there isn't a token record.
Returning `scopes`, `expires_at`, `revoked`, `expired` et. al does not
make sense.

The commit refactors part of `AgnosticTokenRevocationService` to make it
even more generic so that we can specify the presenter used when the API
renders the user. Previously there was an assumption in the code that
there was an object on which we could call `.user`.

This feature exists (still) behind the
`group_agnostic_token_revocation` feature flag. The
rollout issue for that is at
https://gitlab.com/gitlab-org/gitlab/-/issues/463157

Resolves https://gitlab.com/gitlab-org/gitlab/-/issues/468599