Fix timeout when checking group dependencies
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/161215 we
migrated the dependency list empty state to use a new database query.
This query checks a column which is only covered by partial indexes with
a `WHERE archived = false` clause. We forgot to include this scope on
the new query which causes it to have no index coverage so it attempts
to scan the entire table and times out. We only show dependencies from
unarchived projects on the depenency list so the correct behavior is to
only check for existence of unarchived dependencies. This fixes the
timeout as the query can now utilize partial indices.

Changelog: fixed
EE: true