Introduce lock file classes
In https://gitlab.com/gitlab-org/gitlab/-/issues/476177,
our goal is to reproduce the Repository X-Ray functionality
in the monolith.  This is because we can better maintain it
centrally and it will enable us to remove the X-Ray scan
from the CI pipeline.

This is the first step in this process. Here we introduce
lock file classes that contain the logic to parse and
extract libraries for each lock file type.