Add conditional to update isRunningMergeRequestPipeline
This is a temporary UX improvement that will disable
the `Run Pipeline` button until the actual pipeline is created. There
are no error messages surfaced since this is the current UX, but a
failed pipeline will show up if the pipeline is not created successfully.

Changes are all under a feature flag which will be rolled out along with
the BE changes.

Co-authored-by: Laura Montemayor <lmontemayor@gitlab.com>