Add JWT support for internal CI::Minutes APIs
We want to be able to authenticate with a JWT from CustomersDot when
accessing this endpoint. Migrate it to the internal API definition which
has JWT auth, and update the specs to account for this.

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/161707
Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/466368