Make placeholder username more unique by default
When a `source_username` is `nil`, `Gitlab::Utils::Uniquify` will
iterate from 1 to build unique usernames and email addresses.

During the checks of uniqueness queries to `users` .

Previously the only variable in the string uniquify would start with
was the `import_type`, for example `github`. This made it quite likely
that increasingly over time we would iterate a lot to find a unique
username or email address, and issue lots of queries in the process.

This change adds the namespace ID to the starting string to make it more
unique to begin with, as an optimisation to avoid the effect of many
queries to find the uniqueness.