Simplify factory associations with traits when same factory
On an association in a factory, the `factory: [:factory_name, :trait_name]`
syntax is used when we need to use a factory name that's different
from the association name. For example:

association :my_association, factory: [:other_factory, :a_trait]

This will ensure we have an association called my_association,
populated by the factory other_factory, with the trait a_trait.

When the association name is the same as the factory name,
we can change the above line to:

association :my_association, factory: [:my_association, :a_trait]

Which can be further simplified to:

association :my_association, :a_trait

We are doing this very change in this commit.