Ensure discussion reply for linked comments has correct top padding
The discussion reply box for work items was built without the class
`.public-note` applied to it, but that class name was functionally
necessarty to get the top padding necessary to separate the reply from a
linked comment. The style rule can be safely edited to remove this
class.

Changelog: fixed