Fix Git HTTP redirects not working with query parameter
Previously if you attempted to request the following URL:

https://gitlab.com/<repository path>.git?ref=<branch><path>

This would fail with an endless redirect to the origin path.
The previous implementation attempted to strip the `.git` from
the end of a URL, but that meant any query parameters attached
to the URL would prevent that substitution from happening.

Fix this by parsing the URL and stripping the path so that
handle this properly.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/472072

Changelog: fixed