Add published_by to CI Catalog Resource Version
published_by is the user who publishes the CI Catalog Resource Version.
It must be equal to the release author because creating a release
and publishing a version must be done in a single session
via release-cli.

A user can be deleted but it does not mean that we should delete
versions. So, our foreign key strategy here is "nullify".
This is also one of the reason why we require this validation only
`on: :create`.

For old data, we'll implement a background migration to copy
release.author to version.published_by.

Changelog: added