Automatically cancel pipelines in id ascending order
Thus far we've been cancelling Pipelines from this service in
created_at ascending order, and when we cancel a Pipeline with
cascade_to_children: true, we cancel the root first and then proceed
to the downstreams.

We retain the behavior of cancelling the most recent pipelines by
querying IDs in id-descending order.

We then reverse this array and use .order_id_asc on the sequential
.id_in call to cancel the oldest Pipelines first, thus cancelling
child Pipelines after parents.