Add Dangerfile reminder when setting cookies server-side
If a developer sets a cookie from Rails with a `:domain` attribute, they
must remember to un-set this cookie when a user signs out. This can be
hard to remember to do.

This change adds a Danger bot reminder whenever setting cookies is
detected in application code, and also documentation of development
guidelines when working with cookies.