Update Direct Transfer to create placeholder memberships
This Update Directs Transfer's MembersPipeline to create placeholder
memberships when the Import::SourceUser is mapped to a Placeholder User;
in other words, it isn't mapped to a real user. With this
implementation, MembersPipeline will perform the following actions:

1. Create an Import::SourceUser and Placeholder user for every user_id
returned by the member's GraphQL response when there is no corresponding
Import::SourceUser with a source_user_identifier equal to the user_id.
2 Create a membership when the corresponding Import::SourceUser is
mapped to a real user
3. Create a placeholder membership when the corresponding
Import::SourceUser is mapped to a placeholder user