Hide Reset if integration not manually enabled
Some integrations can only be enabled and disabled by the presence of
some other data, for example the GitLab for Slack app integration active
state is toggled when an associated `SlackIntegration` record is created
or destroyed (see `Integrations::SlackInstallation::` classes), and the
GitLab for Jira Cloud app integration when a `JiraConnectSubscription`
record for its inheriting group is created/destroyed (see
`JiraConnectSubscription::CreateService` and `DestroyService`).

In these situations we do not want to allow people to reset the
integration, as resetting destroys the integration, which should only
happen in conjunction with the other data.

This change removes the Reset button from the UI when the integration
cannot be manually enabled and disabled.

https://gitlab.com/gitlab-org/gitlab/-/issues/473128

Changelog: changed