Check only top-level groups on sign-up form endpoint
WHen a user signs up for GitLab, the sign-up form checks to see if the
username they want is already taken. Currently it checks all group names
to see if there is a conflict, however, there is only a conflict for
top-level groups, not sub-groups.

This changes the endpoint to check only top-level groups, not subgroups.

Changelog: fixed