Skip tests for old database migrations
Skip test if migration is older than `Gitlab::Database::MIN_SCHEMA_GITLAB_VERSION`.

This is second attempt to implement this after
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/161544 which had
to be reverted because of the following issues:
- Background migration tests were also skipped
- When old database migration was updated, and its tests skipped,
  `rspec:undercoverage` job was failing.

In this change we address the former by explicitly checking if the
migration being skipped is database migration, and the later by
providing a way to execute the test locally when verifying
`rspec:undercoverage` failures, and updatiing the relevant docs.