Drop work_item_widget_definitions.namespace_id column
Drops the column in preparation of cells 1.0.
Also fixes an incosistency in name casing with the time tracking
widget. New unique name index is case insensitive.

Changelog: other

Remove index reference in down method of migration

A migration references the index
:index_work_item_widget_definitions_on_default_witype_and_name
in the down method of a migration. This doesn't work with
specs