Migrate utility classes in dirs starting with V/W
This is part of the larger ongoing effort to migrate legacy utility
classes to tailwind. We have already migrated thousands of instances in
an automated manner and we came to trust our tooling. So we simply it on
all the JS folders, starting with V and W.

See: https://gitlab.com/groups/gitlab-org/-/epics/13521