Improve Docs for replacing a foreign key with different ON DELETE
The previous documentation recommended to add the new foreign key in
multiple migrations with `validate: false`. There doesn't seem to be any
reason for this. The original discussion tracks back to
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/82414#note_867380679
but this line of reasoning isn't really applicable because the
performance issues being described are mitigated by the fact that
`add_concurrent_foreign_key` already separates validation from creation.

This change also fixes the issue with the original documentation where
our `def down` method was leaving the database schema in a different
state because we weren't validating the foreign key when re-adding it.
This caused confusion at
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/163253#note_2071425164
because this causes the pipeline to fail.

There was also [a discussion on Slack
(internal)](https://gitlab.slack.com/archives/C3NBYFJ6N/p1724384406637759)
and we couldn't think of any reason to separate validation when doing
this process.