Add components_data to Ci Catalog services
Today, when publishing a new version to the CI Catalog, the metadata of
components are fetched from the repository and extracted from the files.

The goal is to move this responsibility to `release-cli` and accept
the required metadata from the client. To this end, a new API endpoint
will accept `components_data` and pass it to
`Ci::Catalog::Resources::ReleaseService`.

In this MR, we are preparing the services to accept `components_data`
before implementing the actual endpoint.