Add limits for import placeholder user creation
During the improved user mapping, placeholder users are created during
an import and assigned as the users of the imported content.

We limit the number of `users` records that can be created during an
import, by imposing limits to the number of placeholder users that can
be created. These limits apply to a root namespace, and take into account
plan and seat count, with more generous limits for higher paid plans and
larger number of paid seats.

When the limit is reached, no more placeholder users are created for the
root namespace. Instead the single "importer user" record that is
associated with their root namespace is assigned as the user of imported
content.

https://gitlab.com/gitlab-org/gitlab/-/issues/455903
https://gitlab.com/gitlab-org/gitlab/-/issues/464401