Use new query to associate SBOM occurrences and vulnerabilities
As part of the [epic][0] to drop the `vulnerability_finding_pipelines`
table, we need to migrate the SBOM ingestion flow off of using the
table.

Currently, during SBOM ingestion, we use the
`vulnerability_finding_pipelines` table to fetch all existing
vulnerability findings associated with a given `pipeline`, and then we
build up an association map in memory mapping every dependency to
existing vulnerability ids, and keeping track of some info, like the
`highest_severity` of the vulnerabilities associated with a given
dependency.

We determined in this [thread][1] that we don't actually have a reason
to scope this query to a given `pipeline`.

This change implements a new query that:

- queries the `vulnerability_occurrences` table directly
- uses the dependency attributes to filter in-database as opposed to in-memory
- doesn't yet remove any of the original code paths to give us any
  easy rollback path

[0]:https://gitlab.com/groups/gitlab-org/-/epics/11241
[1]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/160015#note_2033237030

---

MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/162713
related to: https://gitlab.com/gitlab-org/gitlab/-/issues/472760
Changelog: changed
EE: true