Disable Sidekiq worker open transaction check in JS specs
When transactional tests are enabled, Rails ensures that all threads
use the same connection for accessing the database. However this may
cause a false positive with the Sidekiq worker open transaction
check because:

1. A browser may have multiple requests in-flight.
2. A transaction may start in the middle of another request. For example,
an update to a merge request column initiates a transaction via SAVEPOINT.

To avoid false positives, disable this check for JavaScript feature
specs run. The alternative is to use the deletion strategy, but that
would significantly slow tests.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/478601