Migrate utility classes in files with Auth context
We need special approvals for auth related files, so we migrate all of
those files in one MR in order to reduce the effort needed in other MRs.

This is part of the larger ongoing effort to migrate legacy utility
classes to tailwind. We have already migrated thousands of instances in
an automated manner and we came to trust our tooling.

See also: https://gitlab.com/groups/gitlab-org/-/epics/13521