Backfill published_by_id of catalog_resource_versions
published_by is the user who publishes the CI Catalog Resource Version.
It must be equal to the release author because creating a release
and publishing a version must be done in a single session
via release-cli.

Here, we are backfilling the old data from the releases table by
copying `releases.author` to `catalog_resource_versions.published_by_id`

Changelog: other