GitLab Duo Health Check - Collapsable status
This change introduces some UX improvements
to the GitLab Duo Health Check.

Particulary we now check the health status
async on page load and hide it behind a
collapsable card container.

This should increase performance
of the page load and make it a bit more
intiutive to the user when something is wrong.

Changelog: changed
EE: true