Fix undesired wrapping issue on related items attributes in Safari
In Safari, if a user zooms in while looking at the related MRs widget on
an issuable, the assignee(s) and CI status will get wrapped to separate
lines causing a layout problem.

We don't actually ever want these to wrap as far as I can tell, so the
class `gl-flex-wrap` should be removed.

Changelog: fixed