Optimize ListConfigVariablesService (prefill variables service) (retry)
This is the second step of optimizing `Ci::ListConfigVariablesService`.

Changing parameters in an `calculate_reactive_cache` method is like
changing parameters in a Sidekiq worker. So, we need to follow the same
rules;

1. Add the argument to the worker with a default value (Release M).
2. Add the new argument to all the invocations of the worker
(Release M+1).
3. Remove the default value (Release M+2).

Now, the main reason of this change:

We recently added two changes to this service;

1. https://gitlab.com/gitlab-org/gitlab/-/commit/cd811a5a19ab32e8d25f83c2bb20edd9db31431d
We started passing verify_project_sha to YamlProcessor to check the
passed ref to avoid malicious sha's. (the `ref` parameter accepts
a commit SHA as a value)

2. https://gitlab.com/gitlab-org/gitlab/-/commit/d273ad9544633b1a2a2a168363479402a8c4f749
We started passing the ref name to YamlProcessor when fetching prefill
variables. The reason was this;
When using `$CI_COMMIT_REF_NAME` in `include:rules`, the correct jobs
but the wrong prefilled variables were included because
`$CI_COMMIT_REF_NAME` were not available.

For 1; we first check whether the passed `ref` parameter is actually
a ref in the project. If so, we don't need to verify the calculated
`sha`.

For 2; we don't have to use `RefFinder` because we already have `ref`.

Changelog: other