Rescue TarInvalidError when uploading npm packages
When uploading an NPM package, we fetch its tarball archive and extract the package.json file. We then perform some validations based on the info we extract from package.json file.
However, for some packages, we can end up having an error Gem::Package::TarInvalidError (tar is corrupt, name contains null byte) while trying to find the package.json file in the tarball.
It seems that some files have invalid characters in their names, and the error is raised when calling #full_name
method on each entry in the tarball. In this MR, we rescue this error and call entry.header.name instead to get the entry name.

Changelog: fixed