Set `project_id` when saving pipeline dependency exports
Cells requires all records to have a sharding key. `project_id`,
`group_id`, or `organization_id` can be used as sharding key but
`pipeline_id` cannot, so we need to set the `project_id` for sharding in
this situation. After backfilling, we will add a check constraint to
ensure that one of `project_id`, `group_id`, or `organization_id` is
present and use all three of theses as sharding keys.