Decouple OccurrencesVulnerabilities & vulnerability_finding_pipeline
As part of the [epic][0] to delete the
`vulnerability_finding_pipelines` table, we need to migrate any
application code using that table to a new query

The [`Sbom::Ingestion::Tasks::IngestOccurrencesVulnerabilities`][1]
task ends up using this table when it calls
`occurrence_map.vulnerability_ids`

This commit adds an SQL agg on the `id` column so we can migrate off
of the table in a way that leaves the downstream dependencies
unaffected.

A more detailed description is available in the linked issue.

Note on `occurrence_map.rb` changes
--------------------------------------

There is some fragile code here, where we are creating a new
`@vulnerability_ids` variable and then either running the old code
path or new code path based on the `.nil?` value. This is because we
want to:

1. be able to toggle between old code and new code in production
2. and do that toggling without corrupting any data
3. and we do not have a `project` reference in `occurrence_map` to
   directly control the FF

Alternative implementations exist (e.g. inject a `project`, inject the
FF state, etc.) but this seems fine for how long we expect this code
to stick around (famous last words?)

Note on the `Gitlab::Json.parse` usage
----------------------------------------
In the MR review, a discussion started around ways to avoid manually
parsing the `array_agg` string back to an array in ruby application
code.

Many things were tried (documented in [this thread][2]), though I
couldn't get PostgreSQL to return an integer array directly.

Instead, I wrapped the `array_agg` in a `array_to_json` call in the
SQL query, I then am still parsing the value in ruby. However, the
parsing is now done by `Gitlab::Json.parse`

So,

1. PostgreSQL is handling parsing the agg to json
2. our json library is handling parsing that json back to ruby

So while a roundabout way to get an array, we are at least no longer
introducing any custom parsing code

[0]:https://gitlab.com/groups/gitlab-org/-/epics/11241
[1]:https://gitlab.com/gitlab-org/gitlab/-/blob/master/ee/app/services/sbom/ingestion/tasks/ingest_occurrences_vulnerabilities.rb#L12-19
[2]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/163877#note_2075506549

---

Changelog: changed
EE: true
epic: https://gitlab.com/groups/gitlab-org/-/epics/11241 <br>
resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/480671