Finalize PEP503 sbom component migration
This migration depends on a very specific unique index name to
capture known errors. If the schema changes to add a superset of
this unique index, then the migration will run into not unique
record errors as well, but will not handle them, and eventually
cause the specs to fail. This change prevents a regression, and makes
the test stable by ensuring we don't run the migrations past this
finalized by migration.