Project, group description: highlighted-text style
[0] added a style for the `<mark>` tag used in Markdown. But that style
was not being applied to the description sections in projects and
groups.

In order to reduce duplication, extract that style into into a new
mixin, `str-highlighted`.

Include that mixin from {group, search}.scss, and modify {project,
projects, groups}.scss so that text wrapped in <mark> tags in project
and groups descriptions is rendered consistently.

[0]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/161091

Changelog: fixed