Ensure that settings can be reloaded in tests without changes
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/164290 we
discovered that reloading of the settings caused the
`maximum_text_highlight_size_kilobytes` value to multiply by 1024 each
time. This commit adds a test to ensure that reloading the settings
doesn't change the values to avoid a similar problem in the future.