Backfill `pipeline.project_id` to `dependency_list_export.project_id`
`project_id` will be used as the sharding key for cells for pipeline
dependency list exports. Since pipelines are in a different database,
we need to write a custom migration which executes one query to get the
project ids from the CI database, and then another query to update the
dependency list exports in the main database.

Changelog: changed