Add suffix configuration option to Pipeline execution policies
Uniqueness enforcement causes challenges for compliance teams that may
use the built-in templates twice, causing the jobs to be duplicate.

With this change we are adding a new policy configuration option
suffix: on_conflict/never. The default value is on_conflict.

If policy uses suffix: on_conflict and a job with the same name exists
in the pipeline, we will add suffix in
a format :policy-<security-policy-project-id>-<policy-index>.

The suffix is added following the policy hierarchy.

# Conflicts:
#	ee/app/graphql/resolvers/concerns/construct_security_policies.rb
#	ee/spec/factories/security/policies.rb