Update GitLab Workhorse to AWS SDK for Go v2
AWS SDK for Go v1 has been deprecated
(https://aws.amazon.com/blogs/developer/announcing-end-of-support-for-aws-sdk-for-go-v1-on-july-31-2025/),
and v2 offers a number of improvements:
https://aws.amazon.com/blogs/developer/aws-sdk-for-go-version-2-general-availability/

v2 also allows configuring of AWS_ENDPOINT_URL_STS, which is important
to users of AWS Secret Cloud (not to be confused with GovCloud).

This is behind the workhorse_use_aws_sdk_v2 feature flag, activated by
default.

Note that this commit does not update the S3 tests to use AWS SDK
v1. That should be done as a follow-up.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/482044

Changelog: changed