Receptive Agents: Remove required base64 encoding for certs
This change set removes the base64 encoding that was required for the
certificates in the Receptive Agents API.

This change set DOES NOT warrant a Changelog entry because the feature
is NOT enabled on GitLab.com (via application setting) and is NOT yet
released to self-managed.
Therefore, this is also NOT a breaking change.