Refactor license add-on provisioning
This change is preparation for incoming changes around using more
information from the license's add-on products. At the moment, only the
quantity is used. With the incoming changes, all other info will be
processed and stored within an add-on purchase record as well. To make
the implementation easier, a couple of code places are refactored as
part of this change.