Fix CI_ENVIRONMENT_TIER and CI_ENVIRONMENT_URL variables for seed
When fetching `CI_ENVIRONMENT_TIER` and `CI_ENVIRONMENT_URL`, we look
for a persistent environment in the project with the provided
environment name. We missed this when implementing
`scoped_variables_for_pipeline_seed`.

This MR copies the logic from the model into the variable builder.

I had to move the `persisted_environment` logic from the
`Ci::Deployable` concern to the `Project` model to use it independently
from a `Ci::Build` object in the variables builder.