Introduce class to easier process license add-on purchases
Currently, only the quantity of a license add-on purchase is used. With
incoming changes, more information of a license add-on purchase will be
processed and stored.

This change introduces a new class to easily process a license add-on
purchase. In addition, the mocked license add-on products in various
specs are extended to include additional info.