Make Sidekiq concurrency link more prominent in limited capacity worker
Right now the Sidekiq concurrency limit is more likely to be what people
are searching for if they want to throttle normal Sidekiq workloads so
we should link this at the top of this page to save time. This page for
limited capacity worker is documenting a very specific kind of worker
that is less common so we should make that clear early on.