Stop writing to the `vulnerability_occurrence_pipelines` table
Context
-------------------

We are dropping the `vulnerability_occurrence_pipelines` table in this
[epic][0]

We stopped **reading** from this table by turning the
[`deprecate_vulnerability_occurrence_pipelines`][1]
feature flag on.

The feature flag has globally enabled since `2024-29-08-29` and
everything looks good.

The next steps for the epic are:

1. delete all related application code from codebase
2. add a migration to drop the table

This Commit
-------------------

I had opened a [draft MR][2] that removes all of the
`vulnerability_occurrence_pipelines` application code.

It is quite a large change, and some of the riskier changes were easy
to overlook in the diff

Risk
-------------------

Up till this point, we have kept a clear and easy roll-back path
open.

This commit changes this, as removing the writer-related code will
mean the data in this table is no longer accurate. So this change is
essentially a 1-way door

[0]:https://gitlab.com/groups/gitlab-org/-/epics/11241
[1]:https://gitlab.com/gitlab-org/gitlab/-/issues/450802
[2]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/164111

---

resolves: https://gitlab.com/gitlab-org/gitlab/-/work_items/482623
Related to https://gitlab.com/gitlab-org/gitlab/-/issues/450797
Changelog: changed
EE: true