Fix missing `ci_cd_settings` relation on project creation
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/421050

**Problem**

Relation `ci_cd_settings` is automatically created for every new
project. But it only works if this relation was not touched during
project creation flow.

`group_runners_enabled` field is delegated to `ci_cd_settings`. When
user passes `group_runners_enabled` to the `Projects::CreateService`, it
initializes `ci_cd_settings` relation with a `nil` value. Because of
that, when the project is saved this relation doesn't create a record in
`ci_cd_settings` table.

**Solution**

Explicitly `build` an association for `ci_cd_settings` if it's missing.

Changelog: fixed