Reduce workItemNotesByIid query complexity
The `workItemNotesByIid` query exceeded the maximum complexity for
anonymous (signed out) users, which is set to 200.

The query had a complexity of 218. By reducing the page size, we reduce
that complexity to 180, which gives us some room in the case we need to
query more fields in the future.

Changelog: fixed