Use more info for license add-on provisioning
Prior to this change, only the quantity was used for the add-on
provisioning. However, with the recently updated add-on structure
within the license, all information should be used from the add-on
structure for the provisioning.

This change starts using all provided information of an add-on within
the license to provision it. This way, the created or updated add-on
purchase record will be using more detailed information instead of
using the general information from the license.

With these change, future dated add-ons can be supported in future
changes.

EE: true
Changelog: changed