Fix: Discussions created on legacy Epic are resolvable on WorkItem view
Discussions created on legacy epic were not resolvable on the WorkItem
Epic view, while discussions created on WorkItem Epics are resolvable.

To fix that, we need to ensure that notes with Epic's noteables are
resolvable on the backend.

However, we don't want to add new features on the legacy epic view,
therefore we're hard-coding the `resolvable?` and `resolved?` fields on
Notes payloads on used on the legacy epic view to not render the
`Resolve` buttons there.

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/479180